Auth2

+++ app/Http/Controllers/CommentsApiController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Auth;
 
 class CommentsApiController extends Controller
 {
@@ -15,11 +16,23 @@ class CommentsApiController extends Controller
 
     public function store(){
 
-        return Comment::create([
-            'author_name'=> request('author_name'),
-            'comment_text' => request('comment_text'),
-            'post_id' => request('post_id'),
-        ]);
+        $isGuest = auth()->guest();
+
+        if(! $isGuest){
+            $user_id = auth()->user()->id;
+
+            return Comment::create([
+                'author_name'=> request('author_name'),
+                'comment_text' => request('comment_text'),
+                'post_id' => request('post_id'),
+                'user_id'=> $user_id,
+            ]);
+        }
+        else{
+            return response()->json([
+                "message" => "Unauthorized"
+              ], 401);
+        }
     }
 
     public function getComment($id){
@@ -36,38 +49,69 @@ class CommentsApiController extends Controller
 
     public function update(Request $request, $id){
 
-        if (Comment::where('id', $id)->exists()) {
-            $comment = Comment::find($id);
-            $comment->author_name = is_null($request->author_name) ? $comment->author_name : $request->author_name;
-            $comment->comment_text = is_null($request->comment_text) ? $comment->comment_text : $request->comment_text;
-            $comment->post_id = is_null($request->post_id) ? $comment->post_id : $request->post_id;
-            $comment->save();
+        $isGuest = auth()->guest();
 
-            return response()->json([
-                "message" => "comment updated successfully"
-            ], 200);
-            } else {
-            return response()->json([
-                "message" => "comment not found"
-            ], 404);
+        if(! $isGuest){
 
+            $user_id = auth()->user()->id;
+            $user_role = auth()->user()->role;
+
+
+            if (Comment::where('id', $id)->exists()) {
+                $comment = Comment::find($id);
+
+                if($user_id == $comment->user_id || $user_role == 1){
+
+                $comment->author_name = is_null($request->author_name) ? $comment->author_name : $request->author_name;
+                $comment->comment_text = is_null($request->comment_text) ? $comment->comment_text : $request->comment_text;
+                $comment->post_id = is_null($request->post_id) ? $comment->post_id : $request->post_id;
+                $comment->user_id = $comment->user_id;
+                $comment->save();
+
+                return response()->json([
+                    "message" => "comment updated successfully"
+                ], 200);}
+
+                } else {
+                return response()->json([
+                    "message" => "comment not found"
+                ], 404);
+            }
         }
-
+        else{
+            return response()->json([
+                "message" => "Unauthorized"
+              ], 401);
+        }
     }
 
     public function destroy($id){
 
-        if(Comment::where('id', $id)->exists()) {
-            $comment = Comment::find($id);
-            $comment->delete();
+          $isGuest = auth()->guest();
 
-            return response()->json([
-              "message" => "Comment deleted"
-            ], 202);
-          } else {
-            return response()->json([
-              "message" => "Comment not found"
-            ], 404);
+          if(! $isGuest){
+
+              $user_id = auth()->user()->id;
+              $user_role = auth()->user()->role;
+
+
+              if(Comment::where('id', $id)->exists()) {
+                $comment = Comment::find($id);
+                $comment->delete();
+
+                return response()->json([
+                  "message" => "Comment deleted"
+                ], 202);
+              } else {
+                return response()->json([
+                  "message" => "Comment not found"
+                ], 404);
+              }
+          }
+          else{
+              return response()->json([
+                  "message" => "Unauthorized"
+                ], 401);
           }
     }
 }
